iOS SDK sample app 'theRunAround' modified to only show friends using the app.
Summary:
The app before showed all your friends regardless of if they were using this
app.

With this diff, only friends who are or have used the app will be shown in the
list.

Test Plan:
Run the app with an app id which only some of your friends are using and only
those friends should show in the friends list.

DiffCamp Revision: 169803
Reviewed By: brent
CC: brent
Revert Plan:
OK