Fix the way the SDK interprets UIWebViewDelegate errors
Summary:
Currently if the facebook app encounters a UIWebViewDelegate error, it passes
it to the sdk in the error_code parameter. The SDK should know that if this
parameter is passed, an error occurred, which implies the user *didn't* cancel.
This fixes it.

Test Plan:
disconnected my laptop from the internet, went through the sso flow, verified
the sdk set userDidCancel to NO before calling [self
fbDialogNotLogin:userDidCancel].

DiffCamp Revision: 180310
Reviewed By: brent
CC: brent
Revert Plan:
OK