Add app_id to dialog requests.
Updated the dialog() method to always include app_id. This will prevent
an error if a dialog is requested and the application doesn't have an
access token. App_id must now be specified in the Facebook constructor
as this saves specifying it with every request. Xcode project has been
updated to build with iOS SDK 4.3.