Add the 'localAppId' param to [Facebook authorize]
Summary:
This is the SDK diff to support the local_client_id URL parameter that we added
to the FB app. I named it 'localAppId' to be consistent with the 'appId'
parameter to the same function.

Test Plan:
bound the demo app to 'fb1234foo' and passed 'foo' as the localAppId value.
verified the SDK used the 'fbauth2' url scheme to open the fb app and that the
fb app successfully redirected back to the demo app.

Reviewed By: brent
Reviewers: lshepard, atishm, brent, oyamauchi, jimbru
CC: brent
Differential Revision: 211201