Updated JSON SDK (fixes issue #295)
Sources copied from json-framework, ref 8cabcf40ab1750ea6810818008256dbd744db71c, though Facebook should seriously consider using a git submodule there rather than a copy of the files. As this fix is specifically meant to update the compatibility of the Facebook-iOS-SDK to a recent version of the framework, I didn't feel that it was necessary to "force" Facebook to migrate to a git submodule at the same time, though I would encourage them to do so for obvious reasons.

Thanks to @tinora for reporting this issue, and making the original fixes here which I have rolled into this commit. Hopefully, using a pull request will be a bit easier for Facebook to merge instead of a pasted diff in the issue. ;)