Use the Graph API for the photo uploading example.
Summary:
Edited uploadPhoto to use requestWithGraphPath instead of the legacy REST api
requestWithMethodName method. Also added the publish_stream permission, which
is
needed for uploading photos.

Test Plan:
Run the DemoApp and login. Click the upload button and verify that a photo was
uploaded to the user's Facebook account (pending approval).

Reviewed By: yariv
Reviewers: yariv
CC: yariv
Differential Revision: 257551