Remove auth.expireSession call from SDK
Summary: The expireSession call was put in as a precaution to make sure that
apps sessions were really no longer good. However, this is unnecessary and it
causes issues with apps that were expecting the session to remain usable even
after the user logged out. So let's remove this call and leave it up to the app
developer to expire the session if they want.

Test Plan: Logged in and out of the demo app.

Reviewed By: trvish