Adapted earlier build script to create a .framework and a .pkg.
Summary:
This extends the build_facebook_ios_sdk_static_lib.sh to build
a .framework file, and adds a script to package that and the
templates we will add into an installer .pkg.

These scripts are used by the SDK team to generate binaries
for distribution, and the build_framework script may be
useful to developers who clone the github repository and
want to build the framework with their own local changes.

Also added new headers to the build, and distinguish between
public and project header files.  Since existing projects do not
currently use the .framework, they continue to include headers
directly and are unaffected by public/project.  The legacy
headers are available for copying in the framework download
under FBiOSSDK.framework/Versions/A/LegacyHeaders.

Test Plan:
- Run both scripts
- Verify setup installs a sample template that XCode picks up.
- Verify that the .framework is usable in Vijaye's sample.

Reviewers: jacl, vijaye

Reviewed By: vijaye

CC: gregschechte, security-diffs@lists, yariv

Differential Revision: https://phabricator.fb.com/D431587

Task ID: 981284