Disk Cache and tests
Summary:
This is an SQLite based index and disk cache that keeps the MRU data blobs around, keyed off URL.

I plan to put the cache underneath FBURLConnection, when we identify CDN urls.  This will be useful (in fact, necessary)  when implementing friend picker for people with 1000s of friends - we can't keep all the images we fetch in memory.

Test Plan: Unit tests

Reviewed By: leon

Differential Revision: https://phabricator.fb.com/D439170