Adds a simple login sample, and reflects changes to session object block handling
Summary:
Two changes in this revision: a new sample, tweaks to session behavior

Sample app is a focused sample, which uses only the FBSession SDK class, UIKit and
Foundation. Demonstrates the common login flow for Facebook authentication, including
handling of cached tokens, as well as app-transitions to accomodate authentication.
Note: to avoid sample-bloat, a hyperlink is used to demonstrate logged-in behavior.

Test Plan: manual testing of sample on simulator using iPhone and iPad configurations; session changes to get automated testing in a follow-on revision

Reviewers: mmarucheck, vijaye, leon, caabernathy

Reviewed By: vijaye

CC: yariv

Differential Revision: https://phabricator.fb.com/D441344

Task ID: 958460, 955035