Places picker control
Summary:
A control that allows users to pick a place given location coordinates and a bunch of other graph parameters.  The place pick callback is performed via a delegate, which gets called for every place that is picked.

The contract is that the picked place will be returned as an NSDictioanry.  This would be superseded by the actual Place entity, when we have it.

The pictures are loaded asynchronously and the graph url to CDN url mapping is cached.  The actual image itself is not cached by this control - which is a key goal.  The image will be cached by the underlying FBURLConnection object (not yet done).

Reasonable defaults have been picked for the presentation.

Test Plan: I have a sample app that I used to test this.  I will be polishing that up as an official sample.

Reviewers: leon, jacl, mmarucheck, toddkrabach

Reviewed By: jacl

CC: gregschechte, yariv

Differential Revision: https://phabricator.fb.com/D443097