Reimplements Facebook in terms of Session
Summary:
As part of our migration to the new SDK design, Facebook class is now implemented in terms of
FBSession. Note that the interface to the Facebook class remains consistent, and the implementation
will be migrated further in a follow-on checkin related to FBRequest, and then again down the road
when we update FBDialog.

Test Plan: Manual testing of Hackbook, with whiteboxing all three login modes, token expiration, KVO, etc.

Reviewers: mmarucheck, vijaye, gregschechte, yariv, caabernathy

Reviewed By: mmarucheck

CC: yariv

Differential Revision: https://phabricator.fb.com/D442255

Task ID: 958460