Provide path fallback for package maker
Summary:
When you install package maker from the website, it came down as a DMG and i manually had to install it (and as is the case with other installs, i put it in Applications).

Updating the script to look for the binary in the Applications folder, also.

Test Plan: Verify the script works on my machine

Reviewers: mmarucheck, jacl

Reviewed By: jacl

CC: gregschechte

Differential Revision: https://phabricator.fb.com/D444086