Special case 'false' and 'true' response from graph API.
Summary:
In FBRequestConnection we were not handling 'false' or 'true',
and we were not checking for errors.  This fixes both of these
and passes the results to the user.

False is translated to a graph API error.

True is translated to empty dictionary '{}'.

Also rolled in a fix from Jason to make FBURLConnection.h a project
header rather than public.

Test Plan: Run JustRequestSample with lots of asdf's in both iPhone and iPad.

Reviewers: jacl, vijaye

Differential Revision: https://phabricator.fb.com/D444464

Revert Plan: Safe to revert

Task ID: 1012127