Added SwitchUser sample.
Summary:
Demonstrates an approach to building a Facebook-enabled app that allows switching between
multiple user profiles, for instance on a shared iPad. The sample allows up to 4 profiles
to be added and the user to switch among them. This version of the sample suffers from a
somewhat awkward workflow in that the Facebook iOS app will try to authorize its current
idea of the logged-on user, even if the user of this app wants to add a new user. A workaround
for this is to log out of the Facebook app and log back in as the desired user prior to
adding their profile in the sample. A cleaner workflow will require a mechanism to force
the Facebook app to ignore the currently logged-on user and begin the authorization process
with a login prompt.

Test Plan:
- Ran under iPhone and iPad simulator
- Logged on as multiple test accounts
- Removed some profiles
- Added profiles back
- Verified that "Using" tab reflected currently logged-on user

Reviewers: vijaye, mmarucheck, jacl, gregschechte

Reviewed By: jacl

CC: bhiller

Differential Revision: https://phabricator.fb.com/D450917

Revert Plan: N/A

Task ID: 1015139