Introduces unit testing mode for FBSession
Summary:
Adds a unit testing mode to FBSession, along with a supporting script to configure a
system for unit testing. The workflow is:
* run scripts/configure_simulator_for_unit_tests.sh APP_ID APP_SECRET
* open the facebook-ios-sdk project
* use command-U to run unit tests

In addition to this new workflow for SDK developers, the addition of the static
sessionForUnitTestingWithPermissions to FBSession, allows users to write their own
logic and application unit tests, and use configure_simulator_for_unit_tests.sh to
establish a unit testing routine for their projects, which does not encourage the
hard-coding of application secrets in git repos.

Test Plan: Unit test introduced, run, etc.

Reviewers: mmarucheck, vijaye, clang

CC: gregschechte, platform-diffs@lists, yariv

Differential Revision: https://phabricator.fb.com/D466276

Task ID: 898462