Added Hackbook back into distribution
Summary:
Hackbook previously compiled against all sources in the SDK.

To port a project to the new SDK, users need to:
- (*)Remove all SDK source file references (but not FBConnect.bundle)
- (*)Add file references to SDK/Versions/A/DeprecatedHeaders/*
- (*)Add SDK framework reference.
- Add sqlite3 and quartz framework references.

Items marked with a (*) are part of this change to Hackbook, and are
similar in scope to previous updates and adding the other framework
references.  The goal of zero code change is maintained by this check-in.

Also updated the build script to not remove Hackbook.

Test Plan: Build distribution, install, run Hackbook from installed samples.

Reviewers: caabernathy, clang, gregschechte, jacl

CC: security-diffs@lists

Differential Revision: https://phabricator.fb.com/D460947

Revert Plan: Safe to revert

Task ID: 1012668