[ios-sdk] added Scrumptious sample app
Summary:
The Scrumptious app was floating around via email, etc. We want to use it as an
example of usage of the Facebook iOS SDK. This checkin updates it to use the latest
SDK classes. All functionality should be present except for photo uploading, which
is not yet implemented. Some changes to naming, etc., were undertaken to conform
more closely with iOS community expectations.

This checkin also fixes an issue with FBGraphObjectTableSelection, which was using
object identity to determine whether an object was already selected or not. This did
not work properly across queries, as different FBGraphObjects are created for the
same underlying FB entity. Changed this to look at the ID of graph objects rather than
object identity.

Test Plan:
- Ran the app, picked a place, a meal, and a friend.
- Clicked Announce
- Went to Facebook and verified that a story was posted with the above info.

Reviewers: jacl, mmarucheck, gregschechte, vijaye

Reviewed By: jacl

CC: platform-diffs@lists, yariv

Differential Revision: https://phabricator.fb.com/D469188

Task ID: 1043680