Simple fixup for SwitchUserSample due to different Xcode versions
Summary:
Chris is running Xcode 4.3.1 which doesn't require ##@interface## decls for private methods.  Xcode 4.2.1 does.  This adds these in.

(I think it's wise for not all of us to update to the latest Xcode, as we don't want to require a version of Xcode that we don't take heavy advantage of just because we didn't try on earlier versions.)

Test Plan:
Compile - bask in its success.
Run - bask in even greater glory.

Reviewers: clang

Reviewed By: clang

CC: mmarucheck, jacl

Differential Revision: https://phabricator.fb.com/D465857

Revert Plan: Revert