Have JustRequestSample show all the results it receives
Summary:
This was bugging me:  JustRequestSample supports comma separated list of IDs to do a multi request with.  However, it was only showing the first result.  Fixed that.

Also modified the placeholder text in the text input box to hint that multiple requests can be asked for.

Test Plan: iphone and ipad versions.  Multiple requests and single requests, with existent and non-existent user ids.

Reviewers: mmarucheck, clang, jacl

Reviewed By: jacl

Differential Revision: https://phabricator.fb.com/D465891