Un-set request delegates on teardown
So that, if you are tearing down your Facebook instance, a request can't sneak up on you
and suddenly try to deliver a message to a now dead delegate.

This could also be done by every client saving outstanding FBRequests and making sure to
nil'ing the delegate on them if the delegate is torn down, but that could be a lot of code
and this feels both simpler and correct.