[ios-sdk] Don't cache images in graph objects.
Summary:
FBGraphObjectTableDataSource was caching UIImage instances as properties on
FBGraphObjects. This potentially consumes a great deal of memory and is largely
redundant since the underlying images are already being cached in sqlite.

Removed the caching code. If image is cached, it will be returned immediately.
If it is not, the image in the appropriate table cell will be updated asynchronously
when the image is available.

Also fixes a small error in FBSession that caused it to think it was running as part
of a unit-test when it was not.

Test Plan:
- Built library
- Ran FriendPickerSample
- Verified that images were being displayed correctly in one of two ways:
  - Immediately (synchronously) if they had already been cached
  - Asynchronously as available if they had not been cached
- Closed and re-ran FriendPickerSample to verify that newly-cached images
from previous run were now displaying synchronously.

Reviewers: jacl, mmarucheck, gregschechte, vijaye

Reviewed By: mmarucheck

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D467232

Task ID: 1061654