[ios sdk] Added paging capability to friends and places picker.
Summary:
Graph API responses typically have paging (next/previous) links at the end of
them. We were not following these links, so displayed data was limited to the
amount returned in the first batch of the query. Added the ability to follow
these links to both FBFriendPickerViewController and FBPlacesPickerViewController.

The behavior is somewhat different in each of the controllers:

For Friends, we will keep following the link automagically until we have populated all
friends into the table view. To avoid overly frustrating the user, we attempt to preserve
some visual context in the form of the first visible row, so even as we add new
rows, they have a visual anchor that doesn't make it look like the table is jumping
wildly all over the place.

For Places, we retrieve the next page only when the user scrolls all the way to the
bottom of the table. The spinner indicates more data is being retrieved, and when it
is ready, the list is scrolled to the first item in the new set of data (if the old
last row is still visible, in case the user scrolled back up).

Places picker no longer sorts places by name, but relies on the server to order results.

Test Plan:
- Ran PlacesPickerSample and FriendsPickerSample, verified that behavior works as above.

- Also introduced articially small batch sizes (e.g., limit=10) via code to verify
user experience is not overly painful even with very frequent asynchronous updates to friends list.

Reviewers: mmarucheck, jacl, vijaye, gregschechte

Reviewed By: mmarucheck

CC: ekoneil, platform-diffs@lists, yariv

Differential Revision: https://phabricator.fb.com/D468972

Task ID: 1061665