[ios-sdk] Unit-testing work - add sharing of test users, auto cleanup.
Summary:
Enhances FBTestSession to work in two different modes. The first, Private,
is essentially the same as before -- a test user will be created when
the session is opened, and it will be deleted when it is closed. The
new one, Shared, attempts to re-use existing test users to avoid the
overhead of creation/deletion. When a shared user is created, a hashed
representation of its permissions is stashed in its user name. Later,
we look for a user with the right (hashed) permissions and, if it is not
currently in use by another FBTestSession, that user is returned. This
scheme relies on test writers not doing anything sneaky like changing
permissions on shared users, not deleting shared users during the middle
of a test run, etc. The Private mode allows test authors to not worry about
these issues, at significant run-time overhead due to having to create
and delete them.

FBTests is enhanced with a number of helpers to create FBTestSessions
with the different modes and permissions. FBTests will perform cleanup
of open sessions when a test finishes, to return the shared users, if
any, to the pool of available users.

Fixed a small bug in JSON serialization -- our parser interprets FBIDs
as NSDecimalNumber, but we were not properly re-serializing them to
strings in, e.g., Open Graph Action posts.

Test Plan:
- Ran all unit tests.
- Ran Scrumptious

Reviewers: jacl, mmarucheck, gregschechte

Reviewed By: jacl

CC: platform-diffs@lists, yariv

Differential Revision: https://phabricator.fb.com/D481020