[ios-sdk] Fix occasional crash in FBRequestConnection.
Summary:
In [FBRequestConnection cancel], we first cancel the underlying FBURLConnection.
This cancellation, however, may cause an error handler to be invoked, which
may in turn release the FBRequestConnection. If that was the final retain of the
FBRequestConnection, the object was being dealloc'ed prior to the rest of the
method being executed, causing a crash.

The accepted Objective-C idiom for such cases is to preface the call to the
potentially-dangerous method (in this case, [FBURLConnection cancel]) with:
  [[self retain] autorelease];

This causes the calling object to remain valid until at least the end of the
current method call.

Also cleaned up a case where we were creating an NSError to represent
cancellation, then turning around and wrapping it inside another NSError,
making it unnecessarily complicated for developers to check whether an
error represents a cancellation (which they initiated) or a real error.

Test Plan:
- Ran PlacesPickerSample with the Network Link Conditioner set to simulate
3G network speeds. This reliably triggered the timing necessary to uncover
the bug previously. With this change in place, sample runs correctly with
or without the network speed limited.
- Ran Scrumptious
- Ran unit tests

Reviewers: jacl, mmarucheck, gregschechte

Reviewed By: mmarucheck

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D482439

Task ID: 1093161