[ios-sdk] Handle out-of-range table index requests due to search bar.
Summary:
It appears that, using iOS' built-in search-bar behavior, it will make
requests to tableView:cellForRowAtIndexPath: even if we have indicated
the table has 0 sections and 0 rows (specifically, it will request
the cell at 0,0 when the search bar is initially brought up without any
text in it). In order to handle this, we return a table cell without
any content in it.

Test Plan:
- Ran PlacesPickerSample, clicked in search bar, tried various
queries (e.g., "coffee") at various speeds, observed in debugger
that tableView:cellForRowAtIndexPath: was being called directly
due to changes in search text control, and empty cell was being
returned correctly.

Reviewers: jacl, mmarucheck, gregschechte

Reviewed By: jacl

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D482468

Task ID: 1084389