[ios-sdk] Ensure an access token is always passed for batch requests.
Summary:
Contrary to the docs, the Graph API batch request must contain an access
token for the top-level request, not just the individual sub-requests. It
is used only for app-level auditing of API usage. We use the access token
from the first request in a batch. If all requests have the same access
token, no further work is necessary. Otherwise, we must force each sub-request
to explicitly specify its access token.

It is not allowable for a request that is part of a batch to have a nil session
or to otherwise not have an access token, as there is no way to specify to
the batch API that a given sub-request should NOT inherit the top-level
access token if it does not have an access token of its own.

Added error checking to prevent starting an FBRequestConnection that has
no requests and has no urlRequest set, as it just results in various
exceptions being thrown from within the class.

Test Plan:
- Ran unit tests
- Ran Scrumptious

Reviewers: jacl, mmarucheck, gregschechte

Reviewed By: jacl

CC: harshdeep, selia, trvish

Differential Revision: https://phabricator.fb.com/D482597

Task ID: 1095492