[ios-sdk] build_scrumptious_distribution should build framework; scripts should not redirect logging
Summary:
Script to build Scrumptious had a TODO to build the framework, but it wasn't.

Also, the build server does its own redirection of build script output, and our scripts redirected
almost anything useful to a log file. Removed the redirection, so build scripts send to stdout/stderr
as usual, leaving redirection, if desired, up to the caller. The exception is the progress messages
designed to reassure the viewer that something is happening -- those are being forced to /dev/tty.

Test Plan: - Ran scripts, saw spew

Reviewers: jacl, mmarucheck, gregschechte

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D486364