[ios-sdk] Print progress messages to stderr rather than /dev/tty
Summary:
The scripts being run on the build server were populating the logs with
errors because /dev/tty is not available to them in non-interactive sessions.
It seems the simplest fix is to just print progress messages to stderr. Almost all
of the logspew goes to stdout, except for errors, so just redirecting stdout on
dev boxes gives essentially the same experience as before.

Test Plan: - Ran scripts with redirected stdout, still saw progress messages.

Reviewers: mmarucheck, jacl

Reviewed By: mmarucheck

Differential Revision: https://phabricator.fb.com/D488267