Adds internal-only FBRequestConnection support for basic request caching
Summary:
This checkin prepares for support by widgets for caching and pre-fetching of
results. There are no public API changes here, and the update is tested via
a unit test only. Here are the changes:
  * Adds support to FBRequestConnection similar to FBURLConnection for caching of responses
  * Adds an internal start-method variant that specifies an identity to use for caching
  * Adds a unit test to exercise the variant

Note - we are likely to add general FBRequest caching, which is the primary motivation
for putting cache support at this low a level in the stack.

Test Plan: Added unit test which makes a cacheable request, and then makes the same request confirming it is fetched from cache

Reviewers: clang, mmarucheck

Reviewed By: mmarucheck

CC: yariv

Differential Revision: https://phabricator.fb.com/D488810

Task ID: 1093164