[ios-sdk] Add automated versioning support to SDK.
Summary:
Adds a pre-build Run Script step to the Xcode project to auto-generate an
FBSDKVersion-generated.h header file which contains a #define for a version string
to append to the UserAgent header of FBRequestConnection requests.

This version number is based on the output of the 'git describe' command, or "Unknown"
if no version number could be determined. Since major/minor versions are tagged in git,
if the project is being built from a tagged commit, this will contain just the tag name
(i.e., version number). If the framework is being built from a commit on a repo other than
a tagged one, it will additionally contain the number of commits since the tag and the commit
hash. This can be used by the PHP logging code to determine whether a request is from a
"production" build or from one under development.

Example of the version for a tagged commit: v3.0.0b (signifying a beta build)
Example for a development commit: v3.0.0-144-g4daaf0c (144 commits since a production release)

Test Plan:
- Built the project cleanly both from Xcode and via the build script.
- Verified that both unit tests and sample apps were sending a UserAgent based on the commit.

Reviewers: jacl, mmarucheck, gregschechte

Reviewed By: gregschechte

CC: yariv

Differential Revision: https://phabricator.fb.com/D488586

Task ID: 1111425