Set UserAgent and log "internal request" based RequestConnections.
Summary: #1111515 is about our "paging" requests not carrying the proper user agent header.  This is due to logic in FBRequestConnection that didn't set UserAgent, and didn't invoke the logger, when the "urlRequest" property was explicitly set.

Test Plan: Tested with logging turned on... now seeing the right User Agent in all cases

Reviewers: clang, mmarucheck, jacl

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D488651