[ios-sdk] Replace doc generation with appledoc.
Summary:
appledoc is an open-source documentation generator that generates docs
with a style similar to the Apple documentation and that can easily integrate
with Xcode.

Updated build_documentation.sh to use appledoc if installed, and error if not.

Updated build_distribution.sh to install docset in package/Library/Developer/Shared/...
so it will be installed into user's Xcode docs directory.

Test Plan:
- Blew away build directory and any Facebook dirs under ~/Library/Developer/Shared
- Ran build_distribution.sh
- Installed package
- Ran Xcode
- Verified "Facebook iOS SDK" is shown in Organizer

Reviewers: jacl, mmarucheck, gregschechte, ekoneil

Reviewed By: jacl

CC: caabernathy, platform-diffs@lists, yariv

Differential Revision: https://phabricator.fb.com/D490657

Task ID: 1113291