Behavioral changes for FBProfilePictureView control
Summary:
This is in advance of doing the changes for #1113853 about providing a high res square picture for FBProfilePictureView.

Our Profile Picture view control is messed up in a few ways:

* It is always resizing the profile picture to fit the available view.  In the case of profile pictures, this is almost always wrong due to what it does to image quality.

* It is conflating image size, with "cropping" (square, original).

* It isn't taking advantage of the fact that we can know the right image size to load, and placing that burden on the developer.

Thus, a few changes here:

* As soon as we know the image dimensions, we set the contentMode to ensure the image is never larger than its native size, and that it's never cropped by the view.
* We choose the right image variant to display based upon view size.
* We have a new enum for the "cropping"
* We entirely remove the "size" enumeration.  It's not useful for the developer to specify this.

As such, this control is a lot easier to reason about.  This will also extend well when we have multiple sizes of square images.

This change also adds a bunch of UI control to PictureProfileSample for playing with this.  And updating other controls that were setting the size param.

Test Plan: Scrumptious, SwitchUser, MyProfile, and ProfilePictureSample - both iPad and iPhone.

Reviewers: clang, jacl, mmarucheck

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D491131

Task ID: 1113853