Fixes regression of piggybacking, caused by first of two caching checkins
Summary:
This first caching checkin was naive to the behavior of urlRequest, and resulted
in broken batch URLs. The fix is to predict whether batches are possible, and in
these cases defer calls to urlRequest until the time that we have all of the batch
information in place for proper url creation.

Test Plan: Ran unit tests; clang to checkin and confirm with new regression test

Reviewers: clang, mmarucheck

CC: gregschechte

Differential Revision: https://phabricator.fb.com/D492177

Task ID: 1093164