[ios-sdk] Simplify session access in Scrumptious.
Summary:
Our other samples rely on app delegate to remember the current session, while
Scrumptious had an extra property on the main view controller. Fixed to rely on
the app delegate for current FBSession. To simplify the code, there is still a
small wrapper on the view controller that takes care of getting the app delegate,
since the session is used in a number of different methods.

Test Plan:
- Ran Scrumptious

Revert Plan:

Reviewers: jacl, mmarucheck, ayden, caabernathy

Reviewed By: jacl

Differential Revision: https://phabricator.fb.com/D492223

Task ID: 1107262