Fix cache index update
Summary:
FBCacheIndex applied parameters to the SQL statement in the wrong
order for the updateQuery.  This resulted in keeping an old entry
around that referred to a file that had been deleted.  So on the
next boot, we would check for the file and not find it.  This did
not affect creating new entries, so we did not notice this until
recently.

Test Plan:
- build_all.sh
- in Scrumptious, boot once and wait a while so the cache is
  initialized.  Kill the program, and then start it again.  When
  the entry screen appears, quickly click on "With whom?" to
  see friend picker before the round-trip.

Revert Plan:
Safe

Reviewers: clang, jacl

Reviewed By: clang

Differential Revision: https://phabricator.fb.com/D493317

Task ID: 1121111