[ios-sdk] Don't have ginormous stretched spinner in iOS 4.3.
Summary:
On iOS 4.3, the spinner being displayed in graph object table cells was a
MondoSpinner(tm) the width of the entire display. The world is not yet ready for
MondoSpinner(tm), so it is being replaced with a normal-size spinner. The autoresizing
mask flags apparently are interpreted differently on 4.3 and later iOSes. Fixed to not
autosize the spinner, but rather calculate where it should be placed and explicitly set
its center point.

Also cleaned up some math in FBUtility after discovering the CGRectGetMidX/Y function.

Test Plan:
- Ran place/friend picker samples on 4.3 and 5.0, verified that spinners are all normal size.

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: jacl

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D495067

Task ID: 1123885