[ios-sdk] Better serialiation of NSDate and NSNumber properties.
Summary:
We were not serializing NSDate or NSNumber objects in Graph API requests. Since we convert number-ish
things to NSNumber on deserialization, this was a problem for round-tripping objects. Likewise for NSDate,
which we do not produce on deserialization, but which are a natural thing for developers to want to use
for creating objects with date/timestamps.

Removed adding a place tag to OG Action tests because it depends on developers' unit-test app being approved for
tagging, which is an extra burden and doesn't add much to the value of the test.

Test Plan:
- Ran unit tests under 4.3 and 5.1

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: jacl

CC: platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D496047

Task ID: 1109758