[ios-sdk] Pre-fetch places in Scrumptious, don't refresh location on small changes.
Summary:
When we get a valid location in Scrumptious, we now pre-fetch the cache so the next usage of the
place picker will be fast. We also don't want to get notified of small (<50m) changes in location as
this would tend to kill our use of the cache.

Test Plan:
- Ran Scrumptious

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: jacl

Differential Revision: https://phabricator.fb.com/D497449