[ios-sdk] Make consistent use of YES/NO when dealing with type BOOL
Summary:
We were inconsistent in our use of YES/NO rather than true/false when using type BOOL
rather than bool. This didn't pose a real problem, but we should be consistent in using the
correct constants for the particular datatype in question, and, more importantly, it is not
idiomatic in Objective-C to use true/false for BOOL types.

Test Plan:
- Inspection
- Ran unit tests

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: gregschechte

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D504060