Change FriendPickerSample to be based on a UINavigationController.
Summary:
Update FriendPickerSample as discussed to:

- No longer derive a new view controller for friend picker
- Build the sample as a UINavigationController-based sample
- Support the Cancel/Done UI that fbiphone/wilde use Test Plan:

IMO this is both more straightforward and more akin to expected usage.  However, it's definitely
still a bit tricky primarily due to the setup of the UINavigationController.  I think that's
expected, though, since it's not the job of our sample to make the UI navigation paradigms in
iOS easier to program generally.

I didn't pursue the "modeless" example like what Eddie showed in email.  While the content
control can certainly be put to that use, I really think we need to have these samples be single
purpose/single-use and relatively simple.

Key code to review is the mods in FPAppDelegate.m and FPViewController.m.

It is admittedly a bit odd that the Done/Cancel UI, applied to the friend picker controller, are inserted in the sample app because the sample is the one creating the nav controller.  I think it's correct, though, from what I read on all of this.

Note that I haven't yet made a pass to update to new session convenience methods.  Will do that after an initial review here.

Test Plan: Ran both iPhone and iPad on my box.  Lots of Cancel, Done, friend choosing, etc.

Reviewers: jacl, clang, ekoneil

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D505528

Task ID: 1141720